fix(top-n): bound TrackFilter max_selected to 8 bits and drop debug log
max_selected is packed into the low byte of the wire varint, so any value
above 255 would silently corrupt property_type on both client and relay.
Narrow the Rust field to u8 and validate on the TS constructor so the bad
state can't be constructed. Also removes a per-audio-frame debug log left
over from the VAD hangover work.

+++ libs/moqtail-rs/src/model/parameter/message_parameter.rs
@@ -55,9 +55,11 @@ pub enum MessageParameter {
   /// Top-N track filter (moq-transport PR #1401).
   /// Packed VarInt: value = (property_type << 8) | max_selected.
   /// property_type=0x12 selects speech-activity VAD; max_selected is the N value.
+  /// max_selected is u8 because it occupies only the low 8 bits of the packed value —
+  /// a wider type would silently corrupt property_type on overflow.
   TrackFilter {
     property_type: u64,
-    max_selected: u32,
+    max_selected: u8,
   },
 }
 
@@ -106,7 +108,7 @@ impl MessageParameter {
     Self::NewGroupRequest { group }
   }
 
-  pub fn new_track_filter(property_type: u64, max_selected: u32) -> Self {
+  pub fn new_track_filter(property_type: u64, max_selected: u8) -> Self {
     Self::TrackFilter {
       property_type,
       max_selected,
@@ -274,7 +276,7 @@ impl MessageParameter {
           MessageParameterType::NewGroupRequest => Ok(Self::NewGroupRequest { group: *value }),
           MessageParameterType::TrackFilter => {
             let property_type = value >> 8;
-            let max_selected = (value & 0xFF) as u32;
+            let max_selected = (value & 0xFF) as u8;
             Ok(Self::TrackFilter {
               property_type,
               max_selected,
